requirements.txt
functions-framework==3.*

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
